+++ R/mirt-package.R
@@ -5,7 +5,7 @@
 #' Exploratory and confirmatory models can be estimated with quadrature (EM) or 
 #' stochastic (MHRM) methods. Confirmatory bi-factor and two-tier analyses are available 
 #' for modeling item testlets. Multiple group analysis and mixed effects designs also 
-#' are available for detecting differential item functioning and modelling item and 
+#' are available for detecting differential item functioning and modeling item and 
 #' person covariates.
 #'
 #' Users interested in the most recent version of this package can visit