+++ R/mirtCluster.R
@@ -1,6 +1,6 @@
 #' Define a parallel cluster object to be used in internal functions
 #'
-#' This function defines a object that is placed in a relavent internal enviroment defined in mirt. 
+#' This function defines a object that is placed in a relevant internal environment defined in mirt. 
 #' Internal functions such as \code{calcLogLik}, \code{fscores}, etc, will utilize this object 
 #' automatically to capitalize on parallel
 #' processing architecture. The object defined is a call from \code{parallel::makeCluster()}.